requirements.txt
python-dotenv==1.1.0
requests==2.32.3
streamlit==1.44.1
pydantic==2.11.3
pydantic-settings==2.9.1
